front/requirements.txt
aiogram
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.5.3 